# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartCity OS Design Tools API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /snapshots/{id}/sheets:
    get:
      operationId: getSnapshotSheets
      tags: [snapshots]
      summary: List sheets uploaded for a snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Array of sheet summaries (no PNG bytes)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SheetSummary"
        "404":
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: uploadSnapshotSheets
      tags: [snapshots]
      summary: Upload sheets (multipart) for a snapshot
      description: |
        Multipart upload from the Revit add-in. Body must include a `metadata`
        text field containing JSON-encoded SheetMetadata[] plus per-sheet
        `sheet_{i}_thumb` and `sheet_{i}_full` PNG file fields. Idempotent on
        (snapshotId, sheetNumber) — re-uploading replaces existing rows.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-snapshot-secret
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                metadata:
                  type: string
                  description: JSON-encoded array of SheetMetadata
              required:
                - metadata
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetUploadResponse"
        "401":
          description: Missing or invalid snapshot secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Snapshot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Wrong content-type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sheets/{id}/thumbnail.png:
    get:
      operationId: getSheetThumbnail
      tags: [snapshots]
      summary: Sheet thumbnail PNG (512px)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNG image
          content:
            image/png: {}
        "404":
          description: Sheet not found
  /sheets/{id}/full.png:
    get:
      operationId: getSheetFull
      tags: [snapshots]
      summary: Sheet full-resolution PNG (~2048px)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PNG image
          content:
            image/png: {}
        "404":
          description: Sheet not found
...
components:
  schemas:
...
    SheetSummary:
      type: object
      properties:
        id:
          type: string
        snapshotId:
          type: string
        engagementId:
          type: string
        sheetNumber:
          type: string
        sheetName:
          type: string
        viewCount:
          type: integer
          nullable: true
        revisionNumber:
          type: string
          nullable: true
        revisionDate:
          type: string
          nullable: true
        thumbnailWidth:
          type: integer
        thumbnailHeight:
          type: integer
        fullWidth:
          type: integer
        fullHeight:
          type: integer
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - snapshotId
        - engagementId
        - sheetNumber
        - sheetName
        - viewCount
        - revisionNumber
        - revisionDate
        - thumbnailWidth
        - thumbnailHeight
        - fullWidth
        - fullHeight
        - sortOrder
        - createdAt
    SheetUploadResponse:
      type: object
      properties:
        uploaded:
          type: integer
        skipped:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - uploaded
        - skipped
        - failed
        - errors
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...